requirements.txt
pyrogram==2.0.106
tgcrypto
requests
python-dotenv
pymongo
